Moved list of countries into their own file Moved isNonServiceableCountry into Utils and refactored
