+++ client/components/mma/identity/models.ts
@@ -598,46 +598,3 @@ export const COUNTRIES: CountryValues[] = [
 	{ iso: 'ZW', name: 'Zimbabwe' },
 	{ iso: 'AX', name: 'Åland Islands' },
 ];
-
-export const nonServiceableCountries: string[] = [
-	'Albania',
-	'Åland Islands',
-	'Algeria',
-	'Angola',
-	'Armenia',
-	'Bahrain',
-	'Bangladesh',
-	'Bonaire, Saint Eustatius and Saba',
-	'Cameroon',
-	'Congo (Brazzaville)',
-	'Congo (Kinshasa)',
-	'Curaçao',
-	'Dominican Republic',
-	'Egypt',
-	'French Polynesia',
-	'Gabon',
-	'Georgia',
-	'Guadeloupe',
-	'Kazakhstan',
-	'Kyrgyzstan',
-	'Lesotho',
-	'Mauritania',
-	'Martinique',
-	'Mauritania',
-	'Mauritius',
-	'Moldova',
-	'Monaco',
-	'Morocco',
-	'New Caledonia',
-	'Niger',
-	'Oman',
-	'Pakistan',
-	'Senegal',
-	'Serbia',
-	'Tajikistan',
-	'Tanzania',
-	'Togo',
-	'Tunisia',
-	'Uzbekistan',
-	'Zambia',
-];
