+++ client/components/mma/shared/NonServiceableCountries.ts
@@ -1,0 +1,43 @@
+//Countries cannot purchase or switch to S+
+export const nonServiceableCountries: string[] = [
+	'Albania',
+	'Åland Islands',
+	'Algeria',
+	'Angola',
+	'Armenia',
+	'Bahrain',
+	'Bangladesh',
+	'Bonaire, Saint Eustatius and Saba',
+	'Cameroon',
+	'Congo (Brazzaville)',
+	'Congo (Kinshasa)',
+	'Curaçao',
+	'Dominican Republic',
+	'Egypt',
+	'French Polynesia',
+	'Gabon',
+	'Georgia',
+	'Guadeloupe',
+	'Kazakhstan',
+	'Kyrgyzstan',
+	'Lesotho',
+	'Mauritania',
+	'Martinique',
+	'Mauritania',
+	'Mauritius',
+	'Moldova',
+	'Monaco',
+	'Morocco',
+	'New Caledonia',
+	'Niger',
+	'Oman',
+	'Pakistan',
+	'Senegal',
+	'Serbia',
+	'Tajikistan',
+	'Tanzania',
+	'Togo',
+	'Tunisia',
+	'Uzbekistan',
+	'Zambia',
+];
