+++ client/components/mma/switch/SwitchContainer.tsx
@@ -17,11 +17,13 @@ import {
 	LoadingState,
 	useAsyncLoader,
 } from '../../../utilities/hooks/useAsyncLoader';
-import { createProductDetailFetcher } from '../../../utilities/productUtils';
+import {
+	createProductDetailFetcher,
+	isNonServiceableCountry,
+} from '../../../utilities/productUtils';
 import { getBenefitsThreshold } from '../../../utilities/supporterPlusPricing';
 import { GenericErrorScreen } from '../../shared/GenericErrorScreen';
 import { NAV_LINKS } from '../../shared/nav/NavConfig';
-import { nonServiceableCountries } from '../identity/models';
 import { PageContainer } from '../Page';
 import { JsonResponseHandler } from '../shared/asyncComponents/DefaultApiResponseHandler';
 import { DefaultLoadingView } from '../shared/asyncComponents/DefaultLoadingView';
@@ -127,7 +129,7 @@ const AsyncLoadedSwitchContainer = (props: { isFromApp?: boolean }) => {
 	}
 
 	const contributionToSwitch = data.products.filter(isProduct)[0];
-	if (hasNonServiceableCountry(contributionToSwitch)) {
+	if (isNonServiceableCountry(contributionToSwitch)) {
 		return <Navigate to="/" />;
 	}
 	return (
@@ -175,10 +177,6 @@ const RenderedPage = (props: {
 
 function userIsNavigatingBackFromCompletePage(hasCompleted: boolean) {
 	return hasCompleted && !location.pathname.includes('complete');
-}
-
-function hasNonServiceableCountry(productDetail: ProductDetail) {
-	return nonServiceableCountries.includes(productDetail.billingCountry ?? '');
 }
 
 function getThresholds(
